refactor(Output Generators): Moving output files to a new directory generator in internal (#15)
