+++ internal/core/generator/archive.go
@@ -1,4 +1,4 @@
-package output
+package generator
 
 import (
 	"archive/tar"
@@ -8,44 +8,7 @@ import (
 	"io"
 	"os"
 	"github.com/mdgspace/sysreplicate/internal/domain"
-	// "time"
 )
-
-// backupData structure for tarball creation
-// type BackupData struct {
-// 	Timestamp     time.Time               `json:"timestamp"`
-// 	SystemInfo    SystemInfo              `json:"system_info"`
-// 	EncryptedKeys map[string]EncryptedKey `json:"encrypted_keys"`
-// 	EncryptionKey []byte                  `json:"encryption_key"`
-// }
-
-// type SystemInfo struct {
-// 	Hostname string `json:"hostname"`
-// 	Username string `json:"username"`
-// 	OS       string `json:"os"`
-// }
-
-// type EncryptedKey struct {
-// 	OriginalPath  string `json:"original_path"`
-// 	KeyType       string `json:"key_type"`
-// 	EncryptedData string `json:"encrypted_data"`
-// 	Permissions   uint32 `json:"permissions"`
-// }
-
-// type Dotfile struct {
-// 	Path     string      `json:"path"`
-// 	RealPath string      `json:"real_path"`
-// 	IsDir    bool        `json:"is_dir"`
-// 	IsBinary bool        `json:"is_binary"`
-// 	Mode     os.FileMode `json:"mode"`
-// 	Content  string      `json:"content"` // ignore for the binary files
-// }
-
-// type BackupMetadata struct {
-// 	Timestamp time.Time `json:"timestamp"`
-// 	Hostname  string    `json:"hostname"`
-// 	Files     []Dotfile `json:"files"`
-// }
 
 // create a compressed tarball with the backup data
 func CreateBackupTarball(backupData *domain.BackupData, tarballPath string) error {
