+++ internal/core/generator/metadata.go
@@ -1,12 +1,12 @@
-package output
-
+package generator
+	
 import (
 	"encoding/json"
 	"github.com/mdgspace/sysreplicate/internal/domain"
 )
 
 // BuildSystemJSON creates a well-structured JSON object for the system info and packages.
-func BuildSystemJSON(osType, distro, baseDistro string, packages map[string][]string) ([]byte, error) {
+func GenerateMetadata(osType, distro, baseDistro string, packages map[string][]string) ([]byte, error) {
 
 	// type SystemInfo struct {
 	// 	OS         string              `json:"os"`
