+++ internal/core/generator/scripts.go
@@ -1,4 +1,4 @@
-package output
+package generator
 
 import (
 	"fmt"
